refactor(validator): contain Iroh board adapter
